feat: add unit tests for error routes
